chore: 去除编译警告
用cppcheck去除编译警告。
由于QSignalMapper不推荐使用，重写了PropertyGroup类。

Log:
Task: https://pms.uniontech.com/zentao/task-view-96546.html
Influence: 欢迎界面多屏展示。

+++ dde-osd/kblayoutindicator.cpp
@@ -216,7 +216,7 @@ QPixmap KBLayoutIndicator::generateIconPixmap(const QString &layout, qreal dpr)
     r.setTop(r.top() - 14);
     pa.drawText(r, layout, op);
 
-    QPixmap ret = pix.scaled(QSize(iconSize * dpr, iconSize * dpr),
+    QPixmap ret = pix.scaled(QSize(static_cast<int>(iconSize * dpr), static_cast<int>(iconSize * dpr)),
                              Qt::KeepAspectRatio, Qt::SmoothTransformation);
     ret.setDevicePixelRatio(dpr);
